move receive modal components to separate files
